Major refactor
- Refactor module to extend the appropriate backbone way
- Added CommonJS support
- Just lots of cleanup
