+++ bower.json
@@ -3,6 +3,7 @@
   "dependencies": {
     "jquery": "~2.1.1",
     "backbone": "~1.1.2",
-    "Blob": "*"
+    "Blob": "*",
+    "requirejs": "~2.1.15"
   }
 }
